Cosmos Async Client
Added the necessary dependency in the pom.xml-file for building an asynchronous cosmos client with managed identity. The bean for cosmos client builder in the CosmosSpringConfiguration has been rebuilt to (hopefully) provide a working client using the environment variables for AZURE_CLIENT_ID and COSMOS_DB_ACCOUNT_ENDPOINT in the backend-container.

Added a fallback configuration for local development.